LPS-54318 move to internal package
